+++ src/test/java/es/upm/pproject/prePaidCard/AppTest.java
@@ -24,8 +24,8 @@ public class AppTest {
     
     @Test
     public void test2() {
-    	test.buyCard("Alvaro", 1000, "1111");
-       	Integer amount = 2000;
+    	test.buyCard("Alvaro", (long) 1000, "1111");
+       	Long amount = Long.valueOf(2000);
        	long id = 0;
     	Assertions.assertThrows(NotEnoughMoneyException.class, () -> {test.payCard(id, "1111", amount);});
        	long idExc = 10;
@@ -35,9 +35,9 @@ public class AppTest {
 	
     @Test
     public void test3() throws CardDoesntExistException, ExpiredCardException, NotEnoughMoneyException, WrongPINException {
-    	test.buyCard("Alvaro", 1000, "1111");
-    	Integer amount = 100;
-    	Integer result = 1100;
+    	test.buyCard("Alvaro", (long) 1000, "1111");
+    	Long amount = Long.valueOf(100);
+    	Long result = Long.valueOf(1100);
     	long id = 0;
     	Assertions.assertEquals(result, test.chargeCard(id, "1111", amount));
     	Assertions.assertThrows(WrongPINException.class, () -> {test.chargeCard(id, "111", amount);});
@@ -48,9 +48,9 @@ public class AppTest {
 	
     @Test
     public void test4() throws CardDoesntExistException, WrongPINException {
-    	test.buyCard("Alvaro", 1000, "1111");
+    	test.buyCard("Alvaro", (long) 1000, "1111");
        	long id = 0;
-       	Integer result = 1000;
+       	Long result = Long.valueOf(1000);
     	Assertions.assertEquals(result, test.consultBalance(id, "1111"));
     	Assertions.assertThrows(WrongPINException.class, () -> {test.consultBalance(id, "1234");});
     	long idExc = 10;
@@ -59,7 +59,7 @@ public class AppTest {
     
     @Test
     public void test5() {
-    	test.buyCard("Alvaro", 1000, "1111");
+    	test.buyCard("Alvaro", (long) 1000, "1111");
        	Integer amount = 400;
        	long id = 0;
     	Assertions.assertThrows(WrongPINException.class, () -> {test.payCard(id, "1011", amount);});
@@ -69,11 +69,11 @@ public class AppTest {
     
     @Test
     public void test6() throws CardDoesntExistException, WrongPINException {
-    	test.buyCard("Alvaro", 1000, "1111");
+    	test.buyCard("Alvaro", (long) 1000, "1111");
        	long id = 0;
     	Assertions.assertThrows(WrongPINException.class, () -> {test.changePin(id, "5612", "9999");});
     	test.changePin(id, "1111", "9999");
-    	Integer result = 1000;
+    	Long result = Long.valueOf(1000);
     	Assertions.assertEquals(result, test.consultBalance(id, "9999"));
     	long idExc = 3;
     	Assertions.assertThrows(CardDoesntExistException.class, () -> {test.changePin(idExc, "9999", "1111");});
@@ -82,9 +82,9 @@ public class AppTest {
     
     @Test
     public void test7() throws CardDoesntExistException, WrongPINException, ExpiredCardException, NotEnoughMoneyException {
-    	test.buyCard("Alvaro", 1000, "1111");
+    	test.buyCard("Alvaro", (long) 1000, "1111");
        	long id = 0;
-       	Integer amount = 400;
+       	Long amount = Long.valueOf(400);
     	test.payCard(id, "1111", amount);
     	String result = "";
     	test.consultMovements(id, "1111");
